Better handling of chain removal.
Fix the chain removal bug. Previously, the detection of delete events
was incomplete and chains weren't removed properly.

When the last chain is removed, remove also the "chains" field from the
authservice configuration file. This causes authservice to not accept
the configuration (and be in crash loop), but it is still better than
having an outdated authservice configuration.